Remove compiler warnings, one leak report gone, better debug for VS2017
Unclutter the leak report
..\ctp2_code\ai\mapanalysis\mapgrid.h

Debug version has no ImageHasSafeExceptionHandlers, to allow edit and continue
..\ctp2_code\ctp\civctp.vcxproj

Fixed conversion warnings:
..\ctp2_code\ai\CityManagement\governor.cpp
..\ctp2_code\ai\ctpai.cpp
..\ctp2_code\ai\diplomacy\foreigner.cpp
..\ctp2_code\ai\mapanalysis\mapanalysis.cpp
..\ctp2_code\ai\strategy\agents\agent.cpp
..\ctp2_code\ai\strategy\goals\Goal.cpp
..\ctp2_code\ai\strategy\goals\Goal.h
..\ctp2_code\ai\strategy\scheduler\Scheduler.h
..\ctp2_code\ai\strategy\scheduler\scheduler.cpp
..\ctp2_code\ai\strategy\squads\squad_Strength.h
..\ctp2_code\gs\gameobj\CityData.cpp
..\ctp2_code\gs\gameobj\Happy.cpp
..\ctp2_code\gs\gameobj\tradeutil.cpp
..\ctp2_code\gs\world\cellunitlist.cpp
..\ctp2_code\gs\world\cellunitlist.h

Fixed warnings about deprecated functions:
..\ctp2_code\ctp\civ3_main.cpp
..\ctp2_code\ctp\civapp.cpp
..\ctp2_code\ctp\ctp2_utils\c3cmdline.cpp
..\ctp2_code\ctp\ctp2_utils\c3debug.cpp
..\ctp2_code\ctp\ctp2_utils\c3debug.h
..\ctp2_code\ctp\ctp2_utils\c3files.cpp
..\ctp2_code\ctp\ctp2_utils\c3files.h
..\ctp2_code\ctp\ctp2_utils\tracklen.cpp
..\ctp2_code\ctp\debugtools\debugcallstack.cpp
..\ctp2_code\ctp\debugtools\debugmemory.cpp
..\ctp2_code\ctp\fingerprint\endecode.c
..\ctp2_code\ctp\fingerprint\shroud.c
..\ctp2_code\gfx\spritesys\Anim.cpp
..\ctp2_code\gfx\spritesys\FacedSprite.cpp
..\ctp2_code\gfx\spritesys\GoodSpriteGroup.cpp
..\ctp2_code\gfx\spritesys\Sprite.cpp
..\ctp2_code\gfx\spritesys\SpriteGroup.cpp
..\ctp2_code\gfx\spritesys\UnitSpriteGroup.cpp
..\ctp2_code\gfx\spritesys\effectspritegroup.cpp
..\ctp2_code\gfx\tilesys\tileutils.cpp
..\ctp2_code\gs\database\StrDB.cpp
..\ctp2_code\gs\database\highscoredb.cpp
..\ctp2_code\gs\database\profileDB.cpp
..\ctp2_code\gs\dbgen\Datum.cpp
..\ctp2_code\gs\dbgen\MemberClass.cpp
..\ctp2_code\gs\dbgen\RecordDescription.cpp
..\ctp2_code\gs\dbgen\ctpdb.cpp
..\ctp2_code\gs\dbgen\ctpdb.h
..\ctp2_code\gs\events\GameEventHook.cpp
..\ctp2_code\gs\events\GameEventManager.cpp
..\ctp2_code\gs\fileio\civscenarios.cpp
..\ctp2_code\gs\gameobj\CTP2Combat.cpp
..\ctp2_code\gs\gameobj\CriticalMessagesPrefs.cpp
..\ctp2_code\gs\gameobj\Diplomacy_Log.cpp
..\ctp2_code\gs\gameobj\Player.cpp
..\ctp2_code\gs\gameobj\bldque.cpp
..\ctp2_code\gs\slic\slic.l
..\ctp2_code\gs\slic\slic.y
..\ctp2_code\gs\slic\slicif.cpp
..\ctp2_code\gs\utility\ErrMsg.h
..\ctp2_code\gs\utility\TurnCnt.cpp
..\ctp2_code\gs\world\wldgen.cpp
..\ctp2_code\net\general\net_packet.h
..\ctp2_code\net\io\net_anet.cpp
..\ctp2_code\ui\aui_common\aui_bitmapfont.cpp
..\ctp2_code\ui\aui_ctp2\keymap.cpp
..\ctp2_code\ui\aui_sdl\aui_sdl.cpp
..\ctp2_code\ui\aui_sdl\aui_sdlkeyboard.cpp
..\ctp2_code\ui\aui_sdl\aui_sdlmouse.cpp
..\ctp2_code\ui\aui_sdl\aui_sdlsurface.cpp
..\ctp2_code\ui\interface\EditQueue.cpp
..\ctp2_code\ui\interface\MapCopyBuffer.cpp
..\ctp2_code\ui\interface\backgroundwin.cpp
..\ctp2_code\ui\ldl\ldl.l
..\ctp2_code\ui\ldl\ldl.y
..\ctp2_code\ui\ldl\ldlif.cpp

+++ ctp2_code/gs/gameobj/CityData.cpp
@@ -5178,7 +5178,7 @@ bool CityData::ChangeCurrentlyBuildingItem(sint32 category, sint32 item_type)
 		penalty = std::min<double>
 		                  (
 		                   1.0,
-		                   std::max<float>(0.0, 1.0 - penalty * 0.01)
+		                   std::max<double>(0.0, 1.0 - penalty * 0.01)
 		                  );
 
 		m_shieldstore = static_cast<sint32>(static_cast<double>(m_shieldstore_at_begin_turn) * penalty);
